+++ apps/web/screens/Article/Article.tsx
@@ -51,7 +51,7 @@ import { useI18n } from '@island.is/web/i18n'
 import { withMainLayout } from '@island.is/web/layouts/main'
 import type { Screen } from '@island.is/web/types'
 import { CustomNextError } from '@island.is/web/units/errors'
-import { extractNamespaceFromOrganization } from '@island.is/web/utils/extractCustomTopLoginButtonItemFromOrganization'
+import { extractNamespaceFromOrganization } from '@island.is/web/utils/extractNamespaceFromOrganization'
 import { createNavigation } from '@island.is/web/utils/navigation'
 import { getOrganizationLink } from '@island.is/web/utils/organization'
 import { webRichText } from '@island.is/web/utils/richText'
