+++ apps/web/utils/extractNamespaceFromOrganization.ts
